refactor: ui
-Refactor animation from home page and move it to ui folder.

-Reorganize ui folder to manage purely visual components vs interactive ones.
