+++ app/client/components/ui/interactive/Keyboard.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useMemo, useState } from "react";
-import useHighlightKeys from "../hooks/useHighlightKeys";
-import KeyboardData from "../data/KeyboardData";
-import GenerateDefaultStylingForKeys from "../utils/generators/GenerateDefaultStylingForKeys";
-import useKeyPress from "../hooks/useKeyPress";
-import { useTheme } from "../context/ThemeContext";
-import SimulateKeyPress from "../utils/other/SimulateKeyPress";
-import { useSettings } from "../context/SettingsContext";
+import useHighlightKeys from "../../hooks/useHighlightKeys";
+import KeyboardData from "../../data/KeyboardData";
+import GenerateDefaultStylingForKeys from "../../utils/generators/GenerateDefaultStylingForKeys";
+import useKeyPress from "../../hooks/useKeyPress";
+import { useTheme } from "../../context/ThemeContext";
+import SimulateKeyPress from "../../utils/other/SimulateKeyPress";
+import { useSettings } from "../../context/SettingsContext";
 
 //Theres a lot of object/array manipulation for the initial setup so to improve readability it is going into it's own function
 function DefaultKeyboardSetup() {
