+++ app/root.tsx
@@ -1,7 +1,6 @@
 import "./tailwind.css";
 import NavBar from "./client/components/navigation/NavBar";
 import Footer from "./client/components/navigation/Footer";
-import ReactGA from "react-ga4";
 import cloudflareR2API from "./client/components/api/cloudflareR2API";
 import { ungzip } from "pako";
 
@@ -11,11 +10,9 @@ import {
   Meta,
   Scripts,
   ScrollRestoration,
-  useLocation,
   json,
   ClientLoaderFunctionArgs,
 } from "@remix-run/react";
-import { useEffect, useState } from "react";
 import {
   ThemeProvider,
   useTheme,
@@ -94,6 +91,21 @@ export async function clientLoader({ serverLoader }: ClientLoaderFunctionArgs) {
   }
 }
 
+export function Body({ children }: { children: React.ReactNode }) {
+  const { darkThemeActive } = useTheme();
+
+  return (
+    <body className={`pt-6 ${darkThemeActive && "bg-slate-900"}`}>
+      <NavBar />
+      <div>{children}</div>
+      <ScrollRestoration />
+      <Scripts />
+      <Footer />
+    </body>
+  );
+}
+
+
 // Layout Component for rendering HTML structure
 export function Layout({ children }: { children: React.ReactNode }) {
   return (
@@ -117,47 +129,8 @@ export function Layout({ children }: { children: React.ReactNode }) {
   );
 }
 
-export function Body({ children }: { children: React.ReactNode }) {
-  const { darkThemeActive } = useTheme();
-
-  return (
-    <body className={`pt-6 ${darkThemeActive && "bg-slate-900"}`}>
-      <NavBar />
-      <div>{children}</div>
-      <ScrollRestoration />
-      <Scripts />
-      <Footer />
-    </body>
-  );
-}
-
 // App Component for managing application state
 export default function App() {
-  const [stopGADelayOnStart, setStopGADelayOnStart] = useState(false);
-  const pathname = useLocation().pathname;
-
-  useEffect(() => {
-    const loadGoogleAnalyticsAdsense = async () => {
-      await ReactGA.initialize("G-9C6MJ7Y439"); // Initialize Google Analytics
-
-      // Send page view with a custom path
-      ReactGA.send({
-        hitType: "pageview",
-        page: pathname,
-        title: "Custom Title",
-      });
-    };
-
-    const delay = stopGADelayOnStart ? 0 : 4000;
-
-    setStopGADelayOnStart(true);
-
-    const timer = setTimeout(loadGoogleAnalyticsAdsense, delay);
-
-    return () => clearTimeout(timer);
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pathname]);
 
   return <Outlet />;
 }
