+++ public/sitemap.xml
@@ -2,62 +2,62 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://www.wordskull.com/wildcard</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.795Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/blog</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/cookies-policy</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/faq</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/games</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/privacy-policy</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/settings</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/sitemap</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/stats</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/terms-of-service</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-easy-mode</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-extreme-mode</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-hard-mode</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-medium-mode</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/_index</loc>
-    <lastmod>2024-10-08T07:26:43.889Z</lastmod>
+    <lastmod>2024-10-08T19:56:17.796Z</lastmod>
   </url>
 </urlset>
